Create new google-cloud-storage gem package
