+++ google-cloud-bigquery/acceptance/bigquery/table_test.rb
@@ -191,7 +191,6 @@ describe Google::Cloud::Bigquery::Table, :bigquery do
   end
 
   it "imports data from a file in your bucket" do
-    skip "Skipping because google-cloud-storage isn't available yet"
     begin
       bucket = Google::Cloud.storage.create_bucket "#{prefix}_bucket"
       file = bucket.create_file local_file
@@ -225,7 +224,6 @@ describe Google::Cloud::Bigquery::Table, :bigquery do
   end
 
   it "extracts data to a url in your bucket" do
-    skip "Skipping because google-cloud-storage isn't available yet"
     begin
       # Make sure there is data to extract...
       load_job = table.load local_file
@@ -286,7 +284,6 @@ describe Google::Cloud::Bigquery::Table, :bigquery do
   end
 
   it "extracts data to a file in your bucket" do
-    skip "Skipping because google-cloud-storage isn't available yet"
     begin
       # Make sure there is data to extract...
       load_job = table.load local_file
