+++ google-cloud-bigquery/test/google/cloud/bigquery/table_extract_test.rb
@@ -37,7 +37,6 @@ describe Google::Cloud::Bigquery::Table, :extract, :mock_bigquery do
                                                   bigquery.service }
 
   it "can extract itself to a storage file" do
-    skip "Skipping because google-cloud-storage isn't available yet"
     mock = Minitest::Mock.new
     bigquery.service.mocked_service = mock
     job_gapi = extract_job_gapi(table, extract_file)
@@ -51,7 +50,6 @@ describe Google::Cloud::Bigquery::Table, :extract, :mock_bigquery do
   end
 
   it "can extract itself to a storage url" do
-    skip "Skipping because google-cloud-storage isn't available yet"
     mock = Minitest::Mock.new
     bigquery.service.mocked_service = mock
     job_gapi = extract_job_gapi(table, extract_file)
@@ -65,7 +63,6 @@ describe Google::Cloud::Bigquery::Table, :extract, :mock_bigquery do
   end
 
   it "can extract itself as a dryrun" do
-    skip "Skipping because google-cloud-storage isn't available yet"
     mock = Minitest::Mock.new
     bigquery.service.mocked_service = mock
     job_gapi = extract_job_gapi(table, extract_file)
@@ -80,7 +77,6 @@ describe Google::Cloud::Bigquery::Table, :extract, :mock_bigquery do
   end
 
   it "can extract itself and determine the csv format" do
-    skip "Skipping because google-cloud-storage isn't available yet"
     mock = Minitest::Mock.new
     bigquery.service.mocked_service = mock
     job_gapi = extract_job_gapi(table, extract_file)
@@ -96,7 +92,6 @@ describe Google::Cloud::Bigquery::Table, :extract, :mock_bigquery do
   end
 
   it "can extract itself and specify the csv format" do
-    skip "Skipping because google-cloud-storage isn't available yet"
     mock = Minitest::Mock.new
     bigquery.service.mocked_service = mock
     job_gapi = extract_job_gapi(table, extract_file)
@@ -111,7 +106,6 @@ describe Google::Cloud::Bigquery::Table, :extract, :mock_bigquery do
   end
 
   it "can extract itself and specify the csv format and options" do
-    skip "Skipping because google-cloud-storage isn't available yet"
     mock = Minitest::Mock.new
     bigquery.service.mocked_service = mock
     job_gapi = extract_job_gapi(table, extract_file)
@@ -129,7 +123,6 @@ describe Google::Cloud::Bigquery::Table, :extract, :mock_bigquery do
   end
 
   it "can extract itself and determine the json format" do
-    skip "Skipping because google-cloud-storage isn't available yet"
     mock = Minitest::Mock.new
     bigquery.service.mocked_service = mock
     job_gapi = extract_job_gapi(table, extract_file)
@@ -145,7 +138,6 @@ describe Google::Cloud::Bigquery::Table, :extract, :mock_bigquery do
   end
 
   it "can extract itself and specify the json format" do
-    skip "Skipping because google-cloud-storage isn't available yet"
     mock = Minitest::Mock.new
     bigquery.service.mocked_service = mock
     job_gapi = extract_job_gapi(table, extract_file)
@@ -160,7 +152,6 @@ describe Google::Cloud::Bigquery::Table, :extract, :mock_bigquery do
   end
 
   it "can extract itself and determine the avro format" do
-    skip "Skipping because google-cloud-storage isn't available yet"
     mock = Minitest::Mock.new
     bigquery.service.mocked_service = mock
     job_gapi = extract_job_gapi(table, extract_file)
@@ -176,7 +167,6 @@ describe Google::Cloud::Bigquery::Table, :extract, :mock_bigquery do
   end
 
   it "can extract itself and specify the avro format" do
-    skip "Skipping because google-cloud-storage isn't available yet"
     mock = Minitest::Mock.new
     bigquery.service.mocked_service = mock
     job_gapi = extract_job_gapi(table, extract_file)
