+++ google-cloud-bigquery/test/google/cloud/bigquery/table_load_storage_test.rb
@@ -36,7 +36,6 @@ describe Google::Cloud::Bigquery::Table, :load, :storage, :mock_bigquery do
   end
 
   it "can specify a storage file" do
-    skip "Skipping because google-cloud-storage isn't available yet"
     mock = Minitest::Mock.new
     insert_job = Google::Apis::BigqueryV2::Job.new(
       configuration: Google::Apis::BigqueryV2::JobConfiguration.new(
@@ -55,7 +54,6 @@ describe Google::Cloud::Bigquery::Table, :load, :storage, :mock_bigquery do
   end
 
   it "can specify a storage file with format" do
-    skip "Skipping because google-cloud-storage isn't available yet"
     special_file = storage_file "data.json"
     special_url = special_file.to_gs_url
 
@@ -78,7 +76,6 @@ describe Google::Cloud::Bigquery::Table, :load, :storage, :mock_bigquery do
   end
 
   it "can specify a storage file and derive CSV format" do
-    skip "Skipping because google-cloud-storage isn't available yet"
     special_file = storage_file "data.csv"
     special_url = special_file.to_gs_url
 
@@ -101,7 +98,6 @@ describe Google::Cloud::Bigquery::Table, :load, :storage, :mock_bigquery do
   end
 
   it "can specify a storage file and derive CSV format with CSV options" do
-    skip "Skipping because google-cloud-storage isn't available yet"
     special_file = storage_file "data.csv"
     special_url = special_file.to_gs_url
 
@@ -134,7 +130,6 @@ describe Google::Cloud::Bigquery::Table, :load, :storage, :mock_bigquery do
   end
 
   it "can specify a storage file and derive Avro format" do
-    skip "Skipping because google-cloud-storage isn't available yet"
     special_file = storage_file "data.avro"
     special_url = special_file.to_gs_url
 
@@ -157,7 +152,6 @@ describe Google::Cloud::Bigquery::Table, :load, :storage, :mock_bigquery do
   end
 
   it "can specify a storage file and derive Datastore backup format" do
-    skip "Skipping because google-cloud-storage isn't available yet"
     special_file = storage_file "data.backup_info"
     special_url = special_file.to_gs_url
 
@@ -180,7 +174,6 @@ describe Google::Cloud::Bigquery::Table, :load, :storage, :mock_bigquery do
   end
 
   it "can load a Datastore backup file and specify projection fields" do
-    skip "Skipping because google-cloud-storage isn't available yet"
     special_file = storage_file "data.backup_info"
     special_url = special_file.to_gs_url
     projection_fields = ["first_name"]
@@ -205,7 +198,6 @@ describe Google::Cloud::Bigquery::Table, :load, :storage, :mock_bigquery do
   end
 
   it "can specify a storage url" do
-    skip "Skipping because google-cloud-storage isn't available yet"
     mock = Minitest::Mock.new
     insert_job = Google::Apis::BigqueryV2::Job.new(
       configuration: Google::Apis::BigqueryV2::JobConfiguration.new(
@@ -224,7 +216,6 @@ describe Google::Cloud::Bigquery::Table, :load, :storage, :mock_bigquery do
   end
 
   it "can load itself as a dryrun" do
-    skip "Skipping because google-cloud-storage isn't available yet"
     mock = Minitest::Mock.new
     insert_job = Google::Apis::BigqueryV2::Job.new(
       configuration: Google::Apis::BigqueryV2::JobConfiguration.new(
@@ -243,7 +234,6 @@ describe Google::Cloud::Bigquery::Table, :load, :storage, :mock_bigquery do
   end
 
   it "can load itself with create disposition" do
-    skip "Skipping because google-cloud-storage isn't available yet"
     mock = Minitest::Mock.new
     insert_job = Google::Apis::BigqueryV2::Job.new(
       configuration: Google::Apis::BigqueryV2::JobConfiguration.new(
@@ -263,7 +253,6 @@ describe Google::Cloud::Bigquery::Table, :load, :storage, :mock_bigquery do
   end
 
   it "can load itself with create disposition symbol" do
-    skip "Skipping because google-cloud-storage isn't available yet"
     mock = Minitest::Mock.new
     insert_job = Google::Apis::BigqueryV2::Job.new(
       configuration: Google::Apis::BigqueryV2::JobConfiguration.new(
@@ -283,7 +272,6 @@ describe Google::Cloud::Bigquery::Table, :load, :storage, :mock_bigquery do
   end
 
   it "can load itself with write disposition" do
-    skip "Skipping because google-cloud-storage isn't available yet"
     mock = Minitest::Mock.new
     insert_job = Google::Apis::BigqueryV2::Job.new(
       configuration: Google::Apis::BigqueryV2::JobConfiguration.new(
@@ -303,7 +291,6 @@ describe Google::Cloud::Bigquery::Table, :load, :storage, :mock_bigquery do
   end
 
   it "can load itself with write disposition symbol" do
-    skip "Skipping because google-cloud-storage isn't available yet"
     mock = Minitest::Mock.new
     insert_job = Google::Apis::BigqueryV2::Job.new(
       configuration: Google::Apis::BigqueryV2::JobConfiguration.new(
