+++ google-cloud-bigquery/test/google/cloud/bigquery_test.rb
