+++ google-cloud-core/test/google/cloud/credentials_test.rb
