+++ google-cloud-storage/Gemfile
@@ -4,4 +4,3 @@ gemspec
 
 gem "rake"
 gem "google-cloud-core", path: "../google-cloud-core"
-gem "google-cloud-storage", path: "../google-cloud-storage"
