+++ google-cloud-storage/acceptance/storage/bucket_acl_test.rb
