+++ google-cloud-storage/acceptance/storage/bucket_default_acl_test.rb
