+++ google-cloud-storage/acceptance/storage/bucket_test.rb
