+++ google-cloud-storage/acceptance/storage/file_acl_test.rb
@@ -22,8 +22,8 @@ describe Google::Cloud::Storage::File, :acl, :storage do
   let(:bucket_name) { $bucket_names.first }
 
   let(:files) do
-    { logo: { path: "acceptance/data/CloudPlatform_128px_Retina.png" },
-      big:  { path: "acceptance/data/three-mb-file.tif" } }
+    { logo: { path: "../acceptance/data/CloudPlatform_128px_Retina.png" },
+      big:  { path: "../acceptance/data/three-mb-file.tif" } }
   end
   let(:local_file) { File.new files[:logo][:path] }
