+++ google-cloud-storage/acceptance/storage/files_test.rb
@@ -22,8 +22,8 @@ describe "Storage", :files, :storage do
   let(:bucket_name) { $bucket_names.first }
 
   let(:files) do
-    { logo: { path: "acceptance/data/CloudPlatform_128px_Retina.png" },
-      big:  { path: "acceptance/data/three-mb-file.tif" } }
+    { logo: { path: "../acceptance/data/CloudPlatform_128px_Retina.png" },
+      big:  { path: "../acceptance/data/three-mb-file.tif" } }
   end
   let(:filenames) { ["CloudLogo1", "CloudLogo2", "CloudLogo3"] }
