+++ google-cloud-storage/lib/google/cloud/storage/bucket/acl.rb
