+++ google-cloud-storage/lib/google/cloud/storage/bucket/cors.rb
