+++ google-cloud-storage/lib/google/cloud/storage/errors.rb
