+++ google-cloud-storage/lib/google/cloud/storage/file/list.rb
