+++ google-cloud-storage/lib/google/cloud/storage/project.rb
@@ -13,7 +13,7 @@
 # limitations under the License.
 
 
-require "google/cloud/gce"
+require "google/cloud/core/gce"
 require "google/cloud/storage/errors"
 require "google/cloud/storage/service"
 require "google/cloud/storage/credentials"
@@ -80,8 +80,8 @@ module Google
         # @private Default project.
         def self.default_project
           ENV["STORAGE_PROJECT"] ||
-            ENV["GCLOUD_PROJECT"] ||
             ENV["GOOGLE_CLOUD_PROJECT"] ||
+            ENV["GCLOUD_PROJECT"] ||
             Google::Cloud::GCE.project_id
         end
