+++ google-cloud-storage/test/google/cloud/storage/bucket_acl_test.rb
