+++ google-cloud-storage/test/google/cloud/storage/bucket_test.rb
