+++ google-cloud-storage/test/google/cloud/storage/bucket_update_test.rb
