+++ google-cloud-storage/test/google/cloud/storage/default_acl_test.rb
