+++ google-cloud-storage/test/google/cloud/storage/file_test.rb
