+++ google-cloud-storage/test/google/cloud/storage/file_update_test.rb
