+++ google-cloud-storage/test/google/cloud/storage/project_test.rb
