+++ google-cloud-storage/test/google/cloud/storage/signed_url_test.rb
