+++ google-cloud-storage/test/google/cloud/storage/verifier_test.rb
