+++ google-cloud-storage/test/google/cloud/storage_test.rb
